Update train-ml-bot.py
When running this file with alphabeta or rand as bots there's no problem because there's always a winner.
However, on question 3 we're asked to use a bot trained with ml.Bot, which can get to a situation where no victory is possible, thus causing a loop. 
The change introduces a maximum number of turns in the same way as it is made on the play.py file.